+++ CHANGELOG.md
@@ -1,4 +1,10 @@
 # Change Log
+## v1.1.1 - 2017-02-25
+### Fixed
+- Ocasional wrong color on Total gold balance
+### Added
+- Ability to choose font size
+
 ## v1.0 - 2017-02-23
 ### Release
 - Initial Release